feat:  forgot password and reset password api integration

+++ FastAPI/fastapi_blog/backend/routers/users.py
@@ -337,8 +337,9 @@ async def reset_password(
     reset_token = result.scalars().first()
 
     if not reset_token or reset_token.expires_at.replace(tzinfo=UTC) < datetime.now(UTC):
-        await db.delete(reset_token)
-        await db.commit()
+        if reset_token:
+            await db.delete(reset_token)
+            await db.commit()
 
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
